requirements.txt
python-dotenv
Scrapy
Playwright
Scrapy-playwright